+++ app/src/main/java/com/bnyro/contacts/ui/models/ContactsModel.kt
@@ -9,7 +9,7 @@ import androidx.compose.runtime.setValue
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.bnyro.contacts.obj.ContactData
-import com.bnyro.contacts.util.ContactsHelper
+import com.bnyro.contacts.util.DeviceContactsHelper
 import com.bnyro.contacts.util.PermissionHelper
 import kotlinx.coroutines.launch
 
@@ -20,7 +20,7 @@ class ContactsModel : ViewModel() {
     fun loadContacts(context: Context) {
         if (!PermissionHelper.checkPermissions(context, Manifest.permission.READ_CONTACTS)) return
         viewModelScope.launch {
-            val contactsHelper = ContactsHelper(context)
+            val contactsHelper = DeviceContactsHelper(context)
             contacts = contactsHelper.getContactList()
         }
     }
@@ -29,7 +29,7 @@ class ContactsModel : ViewModel() {
     fun deleteContact(context: Context, contact: ContactData) {
         if (!PermissionHelper.checkPermissions(context, Manifest.permission.READ_CONTACTS)) return
         viewModelScope.launch {
-            val contactsHelper = ContactsHelper(context)
+            val contactsHelper = DeviceContactsHelper(context)
             contactsHelper.deleteContacts(listOf(contact))
             contacts = contacts?.filter { it.contactId != contact.contactId }
         }
@@ -39,7 +39,7 @@ class ContactsModel : ViewModel() {
     fun createContact(context: Context, contact: ContactData) {
         if (!PermissionHelper.checkPermissions(context, Manifest.permission.WRITE_CONTACTS)) return
         viewModelScope.launch {
-            val contactsHelper = ContactsHelper(context)
+            val contactsHelper = DeviceContactsHelper(context)
             contactsHelper.createContact(contact)
             loadContacts(context)
         }
@@ -48,13 +48,8 @@ class ContactsModel : ViewModel() {
     @Suppress("MissingPermission")
     fun updateContact(context: Context, contact: ContactData) {
         if (!PermissionHelper.checkPermissions(context, Manifest.permission.WRITE_CONTACTS)) return
-        val contactsHelper = ContactsHelper(context)
+        val contactsHelper = DeviceContactsHelper(context)
         contactsHelper.deleteContacts(listOf(contact))
         contactsHelper.createContact(contact)
-    }
-
-    fun loadAdvancedContactData(context: Context, contact: ContactData): ContactData {
-        val contactsHelper = ContactsHelper(context)
-        return contactsHelper.loadAdvancedData(contact)
     }
 }
