+++ app/src/main/java/com/bnyro/contacts/util/LocalContactsHelper.kt
@@ -1,0 +1,4 @@
+package com.bnyro.contacts.util
+
+class LocalContactsHelper {
+}
