+++ app/src/main/res/values/strings.xml
@@ -37,6 +37,9 @@
     <string name="assistant">Assistant</string>
     <string name="birthday">Birthday</string>
     <string name="anniversary">Anniversary</string>
+    <!-- Storage locations -->
+    <string name="device">Device</string>
+    <string name="local">Local</string>
     <!-- Import & Export -->
     <string name="import_vcf">Import vCard</string>
     <string name="export_vcf">Export vCard</string>
